+++ src/Colour.php
@@ -4,7 +4,10 @@ declare(strict_types = 1);
 namespace Innmind\Colour;
 
 use Innmind\Colour\Exception\DomainException;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Attempt,
+};
 
 /**
  * @psalm-immutable
@@ -119,10 +122,7 @@ enum Colour
      */
     public static function of(string $colour): RGBA|HSLA|CMYKA
     {
-        return self::maybe($colour)->match(
-            static fn($colour) => $colour,
-            static fn() => throw new DomainException($colour),
-        );
+        return self::attempt($colour)->unwrap();
     }
 
     /**
@@ -132,9 +132,19 @@ enum Colour
      */
     public static function maybe(string $colour): Maybe
     {
-        return RGBA::maybe($colour)
-            ->otherwise(static fn() => HSLA::maybe($colour))
-            ->otherwise(static fn() => CMYKA::maybe($colour));
+        return self::attempt($colour)->maybe();
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Attempt<RGBA|HSLA|CMYKA>
+     */
+    public static function attempt(string $colour): Attempt
+    {
+        return RGBA::attempt($colour)
+            ->recover(static fn() => HSLA::attempt($colour))
+            ->recover(static fn() => CMYKA::attempt($colour));
     }
 
     /**