+++ src/HSLA.php
@@ -7,6 +7,7 @@ use Innmind\Colour\Exception\DomainException;
 use Innmind\Immutable\{
     Str,
     Maybe,
+    Attempt,
 };
 
 /**
@@ -41,10 +42,7 @@ final class HSLA
      */
     public static function of(string $colour): self
     {
-        return self::maybe($colour)->match(
-            static fn($self) => $self,
-            static fn() => throw new DomainException($colour),
-        );
+        return self::attempt($colour)->unwrap();
     }
 
     /**
@@ -54,9 +52,19 @@ final class HSLA
      */
     public static function maybe(string $colour): Maybe
     {
+        return self::attempt($colour)->maybe();
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Attempt<self>
+     */
+    public static function attempt(string $colour): Attempt
+    {
         $colour = Str::of($colour)->trim();
 
-        return self::withAlpha($colour)->otherwise(
+        return self::withAlpha($colour)->recover(
             static fn() => self::withoutAlpha($colour),
         );
     }
@@ -251,9 +259,9 @@ final class HSLA
     /**
      * @psalm-pure
      *
-     * @return Maybe<self>
+     * @return Attempt<self>
      */
-    private static function withAlpha(Str $colour): Maybe
+    private static function withAlpha(Str $colour): Attempt
     {
         $matches = $colour
             ->capture(self::PATTERN_WITH_ALPHA)
@@ -262,39 +270,41 @@ final class HSLA
             ->get('hue')
             ->filter(static fn($hue) => \is_numeric($hue))
             ->map(static fn($hue) => (int) $hue)
-            ->flatMap(static fn($hue) => Hue::of($hue));
+            ->flatMap(static fn($hue) => Hue::of($hue)->maybe());
         $saturation = $matches
             ->get('saturation')
             ->filter(static fn($saturation) => \is_numeric($saturation))
             ->map(static fn($saturation) => (int) $saturation)
-            ->flatMap(static fn($saturation) => Saturation::of($saturation));
+            ->flatMap(static fn($saturation) => Saturation::of($saturation)->maybe());
         $lightness = $matches
             ->get('lightness')
             ->filter(static fn($lightness) => \is_numeric($lightness))
             ->map(static fn($lightness) => (int) $lightness)
-            ->flatMap(static fn($lightness) => Lightness::of($lightness));
+            ->flatMap(static fn($lightness) => Lightness::of($lightness)->maybe());
         $alpha = $matches
             ->get('alpha')
             ->filter(static fn($alpha) => \is_numeric($alpha))
             ->map(static fn($alpha) => (float) $alpha)
-            ->flatMap(static fn($alpha) => Alpha::of($alpha));
+            ->flatMap(static fn($alpha) => Alpha::of($alpha)->maybe());
 
-        return Maybe::all($hue, $saturation, $lightness, $alpha)->map(
-            static fn(Hue $hue, Saturation $saturation, Lightness $lightness, Alpha $alpha) => new self(
-                $hue,
-                $saturation,
-                $lightness,
-                $alpha,
-            ),
-        );
+        return Maybe::all($hue, $saturation, $lightness, $alpha)
+            ->map(
+                static fn(Hue $hue, Saturation $saturation, Lightness $lightness, Alpha $alpha) => new self(
+                    $hue,
+                    $saturation,
+                    $lightness,
+                    $alpha,
+                ),
+            )
+            ->attempt(static fn() => new DomainException($colour->toString()));
     }
 
     /**
      * @psalm-pure
      *
-     * @return Maybe<self>
+     * @return Attempt<self>
      */
-    private static function withoutAlpha(Str $colour): Maybe
+    private static function withoutAlpha(Str $colour): Attempt
     {
         $matches = $colour
             ->capture(self::PATTERN_WITHOUT_ALPHA)
@@ -303,24 +313,26 @@ final class HSLA
             ->get('hue')
             ->filter(static fn($hue) => \is_numeric($hue))
             ->map(static fn($hue) => (int) $hue)
-            ->flatMap(static fn($hue) => Hue::of($hue));
+            ->flatMap(static fn($hue) => Hue::of($hue)->maybe());
         $saturation = $matches
             ->get('saturation')
             ->filter(static fn($saturation) => \is_numeric($saturation))
             ->map(static fn($saturation) => (int) $saturation)
-            ->flatMap(static fn($saturation) => Saturation::of($saturation));
+            ->flatMap(static fn($saturation) => Saturation::of($saturation)->maybe());
         $lightness = $matches
             ->get('lightness')
             ->filter(static fn($lightness) => \is_numeric($lightness))
             ->map(static fn($lightness) => (int) $lightness)
-            ->flatMap(static fn($lightness) => Lightness::of($lightness));
+            ->flatMap(static fn($lightness) => Lightness::of($lightness)->maybe());
 
-        return Maybe::all($hue, $saturation, $lightness)->map(
-            static fn(Hue $hue, Saturation $saturation, Lightness $lightness) => new self(
-                $hue,
-                $saturation,
-                $lightness,
-            ),
-        );
+        return Maybe::all($hue, $saturation, $lightness)
+            ->map(
+                static fn(Hue $hue, Saturation $saturation, Lightness $lightness) => new self(
+                    $hue,
+                    $saturation,
+                    $lightness,
+                ),
+            )
+            ->attempt(static fn() => new DomainException($colour->toString()));
     }
 }